browser-use-demo/requirements.txt
browser-use-sdk>=2.0.0
flask>=3.0
python-dotenv>=1.0
playwright>=1.40
